+++ Assets/Scripts/Data/DataMobileObject.cs
@@ -6,8 +6,8 @@ public class DataMobileObject : DataObjectSpawn
 {
     [SerializeField] protected Transform target =null;
     public Transform _target => this.target;
-    [SerializeField] protected float damage;
-    public float _damage => this.damage;
+    [SerializeField] protected int damage;
+    public int _damage => this.damage;
     protected Category category;
     public Category _category => this.category;
     [SerializeField] protected float flySpeed;
@@ -16,7 +16,7 @@ public class DataMobileObject : DataObjectSpawn
     {
         this.flySpeed = flySpeed;
     }
-    public void SetData(Transform attacker, Transform target, float damage, Category category)
+    public void SetData(Transform attacker, Transform target, int damage, Category category)
     {
         this.damage = damage;
         this.category = category;
